LayoutTests/imported/w3c/web-platform-tests/svg/linking/scripted/a.rel-noreferrer-policy-expected.txt


FAIL No Referrer policy attribute on svg anchor element is applied assert_equals: expected "http://localhost:8800/svg/linking/scripted/a.rel-noreferrer-policy.html" but got ""
